# docs/openapi.json
{
  "info": {
    "description": "Submit and manage workflows",
    "title": "REANA Workflow Controller",
    "version": "0.95.0a7"
  },
  "paths": {
...
    "/api/workflows/{workflow_id_or_name}/close": {
      "post": {
        "consumes": [
          "application/json"
        ],
        "description": "This resource is expecting a workflow to close an interactive session within its workspace.",
        "operationId": "close_interactive_session",
        "parameters": [
          {
            "description": "Required. UUID of workflow owner.",
            "in": "query",
            "name": "user",
            "required": true,
            "type": "string"
          },
          {
            "description": "Required. Workflow UUID or name.",
            "in": "path",
            "name": "workflow_id_or_name",
            "required": true,
            "type": "string"
          }
        ],
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "Request succeeded. The interactive session has been closed.",
            "examples": {
              "application/json": {
                "message": "The interactive session has been closed"
              }
            },
            "schema": {
              "properties": {
                "message": {
                  "type": "string"
                }
              },
              "type": "object"
            }
          },
          "400": {
            "description": "Request failed. The incoming data specification seems malformed.",
            "examples": {
              "application/json": {
                "message": "Malformed request."
              }
            }
          },
          "404": {
            "description": "Request failed. Either User or Workflow does not exist.",
            "examples": {
              "application/json": {
                "message": "Workflow 256b25f4-4cfb-4684-b7a8-73872ef455a1 does not exist"
              }
            }
          },
          "500": {
            "description": "Request failed. Internal controller error."
          }
        },
        "summary": "Close an interactive workflow session."
      }
    },
...
    "/api/workflows/{workflow_id_or_name}/open/{interactive_session_type}": {
      "post": {
        "consumes": [
          "application/json"
        ],
        "description": "This resource is expecting a workflow to start an interactive session within its workspace.",
        "operationId": "open_interactive_session",
        "parameters": [
          {
            "description": "Required. UUID of workflow owner.",
            "in": "query",
            "name": "user",
            "required": true,
            "type": "string"
          },
          {
            "description": "Required. Workflow UUID or name.",
            "in": "path",
            "name": "workflow_id_or_name",
            "required": true,
            "type": "string"
          },
          {
            "description": "Optional. Type of interactive session to use, by default Jupyter Notebook.",
            "in": "path",
            "name": "interactive_session_type",
            "required": true,
            "type": "string"
          },
          {
            "description": "Interactive session configuration.",
            "in": "body",
            "name": "interactive_session_configuration",
            "required": false,
            "schema": {
              "properties": {
                "image": {
                  "description": "Replaces the default Docker image of an interactive session.",
                  "type": "string"
                }
              },
              "type": "object"
            }
          }
        ],
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "Request succeeded. The interactive session has been opened.",
            "examples": {
              "application/json": {
                "path": "/dd4e93cf-e6d0-4714-a601-301ed97eec60"
              }
            },
            "schema": {
              "properties": {
                "path": {
                  "type": "string"
                }
              },
              "type": "object"
            }
          },
          "400": {
            "description": "Request failed. The incoming data specification seems malformed.",
            "examples": {
              "application/json": {
                "message": "Malformed request."
              }
            }
          },
          "404": {
            "description": "Request failed. Either User or Workflow does not exist.",
            "examples": {
              "application/json": {
                "message": "Workflow 256b25f4-4cfb-4684-b7a8-73872ef455a1 does not exist"
              }
            }
          },
          "500": {
            "description": "Request failed. Internal controller error."
          }
        },
        "summary": "Start an interactive session inside the workflow workspace."
      }
    },
...
  },
  "swagger": "2.0"
}